refactor: 🔧 extract Pexels fetch logic to shared utility
Moves the collection-fetching logic out of createPexelsLoader() into a
standalone utility so it can be shared with netlify/functions.
